requirements.txt
easyocr==1.7.1
hydra-core==1.3.2
img2table[==1.2.5
jsonlines==4.0.0
numpy==1.26.1
omegaconf==2.3.0
opencv-python==4.8.1.78
pdf2image==1.16.3
pypdf==3.17.0
pytest==7.4.2
scikit-image==0.22.0
selenium==4.14.0
tabulate==0.9.0
webdriver-manager==4.0.1
